+++ lib/Animal.hpp
@@ -1,4 +1,4 @@
-// updated 11 June 2022
+// updated 9 August 2023
 
 #ifndef ANIMAL_HPP
 #define ANIMAL_HPP
@@ -16,7 +16,7 @@ class Animal : public Entity {
     public:
         explicit Animal(std::string name);
         // attack player, returns damage to be dealt
-        unsigned int attack() noexcept override;
+        double attack() noexcept override;
         // deal damage to the animal, returns true if the animal is dead
         bool attack(double dmg) noexcept override;
         std::vector<std::shared_ptr<Object>> getDrops() noexcept override;
