+++ lib/Enemy.hpp
@@ -1,4 +1,4 @@
-// updated 11 June 2022
+// updated 9 August 2023
 
 #ifndef ENEMY_HPP
 #define ENEMY_HPP
@@ -14,7 +14,7 @@ class Enemy : public Entity {
 
     public:
         explicit Enemy(std::string name);
-        unsigned int attack() noexcept override;
+        double attack() noexcept override;
         // deal damage to the enemy, returns true if the enemy is dead
         bool attack(double dmg) noexcept override;
         std::vector<std::shared_ptr<Object>> getDrops() noexcept override;
